move arc_str tests to api package
