+++ raphtory-api/src/core/storage/arc_str.rs
@@ -108,3 +108,35 @@ impl<'a, O: AsRef<str> + 'a> OptionAsStr<'a> for Option<&'a O> {
         self.map(|s| s.as_ref())
     }
 }
+
+#[cfg(test)]
+mod test_arc_str {
+    use std::sync::Arc;
+    use crate::core::storage::arc_str::{ArcStr, OptionAsStr};
+
+    #[test]
+    fn can_compare_with_str() {
+        let test: ArcStr = "test".into();
+        assert_eq!(test, "test");
+        assert_eq!(test, "test".to_string());
+        assert_eq!(test, Arc::from("test"));
+        assert_eq!(&test, &"test".to_string())
+    }
+
+    #[test]
+    fn test_option_conv() {
+        let test: Option<ArcStr> = Some("test".into());
+
+        let opt_str = test.as_str();
+        assert_eq!(opt_str, Some("test"));
+        
+        let test_ref = test.as_ref();
+        let opt_str = test_ref.as_str();
+        assert_eq!(opt_str, Some("test"));
+
+        let test = Some("test".to_string());
+        let opt_str = test.as_str();
+        assert_eq!(opt_str, Some("test"));
+    }
+
+}
