+++ raphtory/src/core/mod.rs
@@ -886,44 +886,3 @@ mod serde_value_into_prop {
         }
     }
 }
-
-#[cfg(test)]
-mod test_arc_str {
-    use crate::core::Prop;
-    use raphtory_api::core::storage::arc_str::{ArcStr, OptionAsStr};
-    use std::sync::Arc;
-
-    #[test]
-    fn can_compare_with_str() {
-        let test: ArcStr = "test".into();
-        assert_eq!(test, "test");
-        assert_eq!(test, "test".to_string());
-        assert_eq!(test, Arc::from("test"));
-        assert_eq!(&test, &"test".to_string())
-    }
-
-    #[test]
-    fn test_option_conv() {
-        let test: Option<ArcStr> = Some("test".into());
-
-        let test_ref = test.as_ref();
-
-        let opt_str = test.as_str();
-        let opt_str3 = test_ref.as_str();
-
-        let test2 = Some("test".to_string());
-        let opt_str_2 = test2.as_str();
-
-        assert_eq!(opt_str, Some("test"));
-        assert_eq!(opt_str_2, Some("test"));
-        assert_eq!(opt_str3, Some("test"));
-    }
-
-    #[test]
-    fn test_prop_min_max() {
-        let v1 = Prop::I64(4);
-        let v2 = Prop::I64(2);
-        assert_eq!(v1.clone().max(v2.clone()), Some(Prop::I64(4)));
-        assert_eq!((v1.min(v2)), Some(Prop::I64(2)));
-    }
-}
